LayoutTests/http/tests/from-origin/fetch-data-iframe-from-origin-same-blocked-expected.txt

CONSOLE MESSAGE: Cancelled load because it violates the resource's From-Origin response header.
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8000/from-origin/resources/fetch.php?fromOrigin=same due to access control checks.
Tests that a cross-origin fetch in a data: iframe load fails if the server blocks cross-origin loads with a 'From-Origin: same' response header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 127.0.0.1 fetch in data: iframe blocked.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
The iframe making a 127.0.0.1 fetch call.
